+++ modules/parallel/pyext/src/worker_handler.py
@@ -4,20 +4,20 @@ import traceback
 import socket
 import threading
 from IMP.parallel import NetworkError
-from IMP.parallel.master_communicator import MasterCommunicator
+from IMP.parallel.manager_communicator import ManagerCommunicator
 from IMP.parallel.util import _TaskWrapper, _ContextWrapper
-from IMP.parallel.util import _ErrorWrapper, _HeartBeat, _SlaveAction
+from IMP.parallel.util import _ErrorWrapper, _HeartBeat, _WorkerAction
 
 
 class _HeartBeatThread(threading.Thread):
 
-    """Periodically send a 'heartbeat' back to the master, so that it can
+    """Periodically send a 'heartbeat' back to the manager, so that it can
        distinguish between failed nodes and long calculations"""
     timeout = 300
 
-    def __init__(self, master):
+    def __init__(self, manager):
         threading.Thread.__init__(self)
-        self._master = master
+        self._manager = manager
         self._event = threading.Event()
 
     def cancel(self):
@@ -30,40 +30,40 @@ class _HeartBeatThread(threading.Thread):
             if self._event.isSet():
                 break
             else:
-                self._master._send(_HeartBeat())
+                self._manager._send(_HeartBeat())
 
 
-class SlaveHandler(object):
+class WorkerHandler(object):
 
-    def __init__(self, master_addr):
-        self._master_addr = master_addr
+    def __init__(self, manager_addr):
+        self._manager_addr = manager_addr
 
     def run(self):
-        print("Connect back to master at %s:%d with ID %s" \
-              % tuple(self._master_addr))
+        print("Connect back to manager at %s:%d with ID %s" \
+              % tuple(self._manager_addr))
         lock = threading.Lock()
-        master = MasterCommunicator(self._master_addr, lock)
-        hb = _HeartBeatThread(master)
+        manager = ManagerCommunicator(self._manager_addr, lock)
+        hb = _HeartBeatThread(manager)
         hb.start()
         try:
-            self._handle_network_io(master)
+            self._handle_network_io(manager)
         finally:
             hb.cancel()
 
-    def _send_exception_to_master(self, master, exc):
+    def _send_exception_to_manager(self, manager, exc):
         try:
             exc_type, exc_value, tb = sys.exc_info()
-            master._send(_ErrorWrapper(exc,
-                                       traceback.format_exception(exc_type, exc_value, tb)))
+            manager._send(_ErrorWrapper(
+                exc, traceback.format_exception(exc_type, exc_value, tb)))
         except socket.error:
-            # ignore errors encountered while trying to send error to master
+            # ignore errors encountered while trying to send error to manager
             pass
 
-    def _handle_network_io(self, master):
+    def _handle_network_io(self, manager):
         setup_args = ()
         while True:
             try:
-                obj = master._recv()
+                obj = manager._recv()
             except NetworkError:
                 break
             try:
@@ -73,19 +73,19 @@ class SlaveHandler(object):
                     else:
                         setup_args = obj.obj()
                 elif isinstance(obj, _TaskWrapper):
-                    master._send(obj.obj(*setup_args))
-                elif isinstance(obj, _SlaveAction):
+                    manager._send(obj.obj(*setup_args))
+                elif isinstance(obj, _WorkerAction):
                     obj.execute()
             except NetworkError:
                 raise
             except Exception as detail:
-                # Send all other exceptions back to the master and reraise
-                self._send_exception_to_master(master, detail)
+                # Send all other exceptions back to the manager and reraise
+                self._send_exception_to_manager(manager, detail)
                 raise
 
 
 def main():
-    h = SlaveHandler([sys.argv[-3], int(sys.argv[-2]), sys.argv[-1]])
+    h = WorkerHandler([sys.argv[-3], int(sys.argv[-2]), sys.argv[-1]])
     h.run()
 
 if __name__ == '__main__':