+++ modules/parallel/test/test_startup.py
@@ -10,9 +10,9 @@ class Tests(IMP.test.TestCase):
     """Test startup of parallel jobs"""
 
     def test_simple(self):
-        """Test that slave tasks can start up and communicate"""
+        """Test that worker tasks can start up and communicate"""
         m = _util.Manager(output='simple%d.out')
-        m.add_slave(IMP.parallel.LocalSlave())
+        m.add_worker(IMP.parallel.LocalWorker())
         c = m.get_context()
         for i in range(10):
             c.add_task(_tasks.SimpleTask(i))
@@ -23,7 +23,7 @@ class Tests(IMP.test.TestCase):
     def test_startup(self):
         """Test context startup callable"""
         m = _util.Manager(output='startup%d.out')
-        m.add_slave(IMP.parallel.LocalSlave())
+        m.add_worker(IMP.parallel.LocalWorker())
         c = m.get_context(startup=_tasks.SimpleTask(("foo", "bar")))
         c.add_task(_tasks.simple_func)
         c.add_task(_tasks.simple_func)
@@ -38,19 +38,19 @@ class Tests(IMP.test.TestCase):
         m = _util.Manager(python="/path/does/not/exist",
                           output='heartbeat%d.out')
         m.heartbeat_timeout = 0.1
-        m.add_slave(IMP.parallel.LocalSlave())
+        m.add_worker(IMP.parallel.LocalWorker())
         c = m.get_context()
         c.add_task(empty_task)
         self.assertRaises(IMP.parallel.NetworkError, list,
                           c.get_results_unordered())
         _util.unlink('heartbeat0.out')
 
-    def test_startup_no_slaves(self):
-        """Test that startup with no slaves causes a failure"""
+    def test_startup_no_workers(self):
+        """Test that startup with no workers causes a failure"""
         m = _util.Manager()
         c = m.get_context()
         c.add_task(_tasks.simple_func)
-        self.assertRaises(IMP.parallel.NoMoreSlavesError, list,
+        self.assertRaises(IMP.parallel.NoMoreWorkersError, list,
                           c.get_results_unordered())